@@ -1,5 +1,5 @@
 module file-transfer
 
-go 1.19
+go 1.22
 
 require google.golang.org/protobuf v1.28.1
