@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.28.1
-// 	protoc        v3.21.12
+// 	protoc-gen-go v1.36.11
+// 	protoc        v6.30.2
 // source: messages.proto
 
 package messages
@@ -11,6 +11,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -21,21 +22,18 @@ const (
 )
 
 type StorageRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	FileName      string                 `protobuf:"bytes,1,opt,name=file_name,json=fileName,proto3" json:"file_name,omitempty"`
+	Size          uint64                 `protobuf:"varint,2,opt,name=size,proto3" json:"size,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	FileName string `protobuf:"bytes,1,opt,name=file_name,json=fileName,proto3" json:"file_name,omitempty"`
-	Size     uint64 `protobuf:"varint,2,opt,name=size,proto3" json:"size,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *StorageRequest) Reset() {
 	*x = StorageRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_messages_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_messages_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *StorageRequest) String() string {
@@ -46,7 +44,7 @@ func (*StorageRequest) ProtoMessage() {}
 
 func (x *StorageRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_messages_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -76,20 +74,17 @@ func (x *StorageRequest) GetSize() uint64 {
 }
 
 type RetrievalRequest struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	FileName      string                 `protobuf:"bytes,1,opt,name=file_name,json=fileName,proto3" json:"file_name,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	FileName string `protobuf:"bytes,1,opt,name=file_name,json=fileName,proto3" json:"file_name,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *RetrievalRequest) Reset() {
 	*x = RetrievalRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_messages_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_messages_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RetrievalRequest) String() string {
@@ -100,7 +95,7 @@ func (*RetrievalRequest) ProtoMessage() {}
 
 func (x *RetrievalRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_messages_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -123,20 +118,17 @@ func (x *RetrievalRequest) GetFileName() string {
 }
 
 type ChecksumVerification struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Checksum      []byte                 `protobuf:"bytes,1,opt,name=checksum,proto3" json:"checksum,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Checksum []byte `protobuf:"bytes,1,opt,name=checksum,proto3" json:"checksum,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *ChecksumVerification) Reset() {
 	*x = ChecksumVerification{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_messages_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_messages_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *ChecksumVerification) String() string {
@@ -147,7 +139,7 @@ func (*ChecksumVerification) ProtoMessage() {}
 
 func (x *ChecksumVerification) ProtoReflect() protoreflect.Message {
 	mi := &file_messages_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -170,21 +162,18 @@ func (x *ChecksumVerification) GetChecksum() []byte {
 }
 
 type Response struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Ok            bool                   `protobuf:"varint,1,opt,name=ok,proto3" json:"ok,omitempty"`
+	Message       string                 `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Ok      bool   `protobuf:"varint,1,opt,name=ok,proto3" json:"ok,omitempty"`
-	Message string `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Response) Reset() {
 	*x = Response{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_messages_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_messages_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Response) String() string {
@@ -195,7 +184,7 @@ func (*Response) ProtoMessage() {}
 
 func (x *Response) ProtoReflect() protoreflect.Message {
 	mi := &file_messages_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -225,21 +214,18 @@ func (x *Response) GetMessage() string {
 }
 
 type RetrievalResponse struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Resp          *Response              `protobuf:"bytes,1,opt,name=resp,proto3" json:"resp,omitempty"`
+	Size          uint64                 `protobuf:"varint,2,opt,name=size,proto3" json:"size,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Resp *Response `protobuf:"bytes,1,opt,name=resp,proto3" json:"resp,omitempty"`
-	Size uint64    `protobuf:"varint,2,opt,name=size,proto3" json:"size,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *RetrievalResponse) Reset() {
 	*x = RetrievalResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_messages_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_messages_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RetrievalResponse) String() string {
@@ -250,7 +236,7 @@ func (*RetrievalResponse) ProtoMessage() {}
 
 func (x *RetrievalResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_messages_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -280,27 +266,24 @@ func (x *RetrievalResponse) GetSize() uint64 {
 }
 
 type Wrapper struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
-	// Types that are assignable to Msg:
+	state protoimpl.MessageState `protogen:"open.v1"`
+	// Types that are valid to be assigned to Msg:
 	//
 	//	*Wrapper_Response
 	//	*Wrapper_StorageReq
 	//	*Wrapper_RetrievalReq
 	//	*Wrapper_RetrievalResp
 	//	*Wrapper_Checksum
-	Msg isWrapper_Msg `protobuf_oneof:"msg"`
+	Msg           isWrapper_Msg `protobuf_oneof:"msg"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Wrapper) Reset() {
 	*x = Wrapper{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_messages_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_messages_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Wrapper) String() string {
@@ -311,7 +294,7 @@ func (*Wrapper) ProtoMessage() {}
 
 func (x *Wrapper) ProtoReflect() protoreflect.Message {
 	mi := &file_messages_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -326,44 +309,54 @@ func (*Wrapper) Descriptor() ([]byte, []int) {
 	return file_messages_proto_rawDescGZIP(), []int{5}
 }
 
-func (m *Wrapper) GetMsg() isWrapper_Msg {
-	if m != nil {
-		return m.Msg
+func (x *Wrapper) GetMsg() isWrapper_Msg {
+	if x != nil {
+		return x.Msg
 	}
 	return nil
 }
 
 func (x *Wrapper) GetResponse() *Response {
-	if x, ok := x.GetMsg().(*Wrapper_Response); ok {
-		return x.Response
+	if x != nil {
+		if x, ok := x.Msg.(*Wrapper_Response); ok {
+			return x.Response
+		}
 	}
 	return nil
 }
 
 func (x *Wrapper) GetStorageReq() *StorageRequest {
-	if x, ok := x.GetMsg().(*Wrapper_StorageReq); ok {
-		return x.StorageReq
+	if x != nil {
+		if x, ok := x.Msg.(*Wrapper_StorageReq); ok {
+			return x.StorageReq
+		}
 	}
 	return nil
 }
 
 func (x *Wrapper) GetRetrievalReq() *RetrievalRequest {
-	if x, ok := x.GetMsg().(*Wrapper_RetrievalReq); ok {
-		return x.RetrievalReq
+	if x != nil {
+		if x, ok := x.Msg.(*Wrapper_RetrievalReq); ok {
+			return x.RetrievalReq
+		}
 	}
 	return nil
 }
 
 func (x *Wrapper) GetRetrievalResp() *RetrievalResponse {
-	if x, ok := x.GetMsg().(*Wrapper_RetrievalResp); ok {
-		return x.RetrievalResp
+	if x != nil {
+		if x, ok := x.Msg.(*Wrapper_RetrievalResp); ok {
+			return x.RetrievalResp
+		}
 	}
 	return nil
 }
 
 func (x *Wrapper) GetChecksum() *ChecksumVerification {
-	if x, ok := x.GetMsg().(*Wrapper_Checksum); ok {
-		return x.Checksum
+	if x != nil {
+		if x, ok := x.Msg.(*Wrapper_Checksum); ok {
+			return x.Checksum
+		}
 	}
 	return nil
 }
@@ -404,62 +397,46 @@ func (*Wrapper_Checksum) isWrapper_Msg() {}
 
 var File_messages_proto protoreflect.FileDescriptor
 
-var file_messages_proto_rawDesc = []byte{
-	0x0a, 0x0e, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x22, 0x41, 0x0a, 0x0e, 0x53, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65,
-	0x73, 0x74, 0x12, 0x1b, 0x0a, 0x09, 0x66, 0x69, 0x6c, 0x65, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x66, 0x69, 0x6c, 0x65, 0x4e, 0x61, 0x6d, 0x65, 0x12,
-	0x12, 0x0a, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04, 0x52, 0x04, 0x73,
-	0x69, 0x7a, 0x65, 0x22, 0x2f, 0x0a, 0x10, 0x52, 0x65, 0x74, 0x72, 0x69, 0x65, 0x76, 0x61, 0x6c,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12, 0x1b, 0x0a, 0x09, 0x66, 0x69, 0x6c, 0x65, 0x5f,
-	0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x66, 0x69, 0x6c, 0x65,
-	0x4e, 0x61, 0x6d, 0x65, 0x22, 0x32, 0x0a, 0x14, 0x43, 0x68, 0x65, 0x63, 0x6b, 0x73, 0x75, 0x6d,
-	0x56, 0x65, 0x72, 0x69, 0x66, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x1a, 0x0a, 0x08,
-	0x63, 0x68, 0x65, 0x63, 0x6b, 0x73, 0x75, 0x6d, 0x18, 0x01, 0x20, 0x01, 0x28, 0x0c, 0x52, 0x08,
-	0x63, 0x68, 0x65, 0x63, 0x6b, 0x73, 0x75, 0x6d, 0x22, 0x34, 0x0a, 0x08, 0x52, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x65, 0x12, 0x0e, 0x0a, 0x02, 0x6f, 0x6b, 0x18, 0x01, 0x20, 0x01, 0x28, 0x08,
-	0x52, 0x02, 0x6f, 0x6b, 0x12, 0x18, 0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18,
-	0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x22, 0x46,
-	0x0a, 0x11, 0x52, 0x65, 0x74, 0x72, 0x69, 0x65, 0x76, 0x61, 0x6c, 0x52, 0x65, 0x73, 0x70, 0x6f,
-	0x6e, 0x73, 0x65, 0x12, 0x1d, 0x0a, 0x04, 0x72, 0x65, 0x73, 0x70, 0x18, 0x01, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x09, 0x2e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x52, 0x04, 0x72, 0x65,
-	0x73, 0x70, 0x12, 0x12, 0x0a, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x04,
-	0x52, 0x04, 0x73, 0x69, 0x7a, 0x65, 0x22, 0x99, 0x02, 0x0a, 0x07, 0x57, 0x72, 0x61, 0x70, 0x70,
-	0x65, 0x72, 0x12, 0x27, 0x0a, 0x08, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x09, 0x2e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x48,
-	0x00, 0x52, 0x08, 0x72, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x32, 0x0a, 0x0b, 0x73,
-	0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x5f, 0x72, 0x65, 0x71, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x0f, 0x2e, 0x53, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
-	0x74, 0x48, 0x00, 0x52, 0x0a, 0x73, 0x74, 0x6f, 0x72, 0x61, 0x67, 0x65, 0x52, 0x65, 0x71, 0x12,
-	0x38, 0x0a, 0x0d, 0x72, 0x65, 0x74, 0x72, 0x69, 0x65, 0x76, 0x61, 0x6c, 0x5f, 0x72, 0x65, 0x71,
-	0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x11, 0x2e, 0x52, 0x65, 0x74, 0x72, 0x69, 0x65, 0x76,
-	0x61, 0x6c, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x48, 0x00, 0x52, 0x0c, 0x72, 0x65, 0x74,
-	0x72, 0x69, 0x65, 0x76, 0x61, 0x6c, 0x52, 0x65, 0x71, 0x12, 0x3b, 0x0a, 0x0e, 0x72, 0x65, 0x74,
-	0x72, 0x69, 0x65, 0x76, 0x61, 0x6c, 0x5f, 0x72, 0x65, 0x73, 0x70, 0x18, 0x04, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x12, 0x2e, 0x52, 0x65, 0x74, 0x72, 0x69, 0x65, 0x76, 0x61, 0x6c, 0x52, 0x65, 0x73,
-	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x48, 0x00, 0x52, 0x0d, 0x72, 0x65, 0x74, 0x72, 0x69, 0x65, 0x76,
-	0x61, 0x6c, 0x52, 0x65, 0x73, 0x70, 0x12, 0x33, 0x0a, 0x08, 0x63, 0x68, 0x65, 0x63, 0x6b, 0x73,
-	0x75, 0x6d, 0x18, 0x05, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x15, 0x2e, 0x43, 0x68, 0x65, 0x63, 0x6b,
-	0x73, 0x75, 0x6d, 0x56, 0x65, 0x72, 0x69, 0x66, 0x69, 0x63, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x48,
-	0x00, 0x52, 0x08, 0x63, 0x68, 0x65, 0x63, 0x6b, 0x73, 0x75, 0x6d, 0x42, 0x05, 0x0a, 0x03, 0x6d,
-	0x73, 0x67, 0x42, 0x0c, 0x5a, 0x0a, 0x2e, 0x2f, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x73,
-	0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_messages_proto_rawDesc = "" +
+	"\n" +
+	"\x0emessages.proto\"A\n" +
+	"\x0eStorageRequest\x12\x1b\n" +
+	"\tfile_name\x18\x01 \x01(\tR\bfileName\x12\x12\n" +
+	"\x04size\x18\x02 \x01(\x04R\x04size\"/\n" +
+	"\x10RetrievalRequest\x12\x1b\n" +
+	"\tfile_name\x18\x01 \x01(\tR\bfileName\"2\n" +
+	"\x14ChecksumVerification\x12\x1a\n" +
+	"\bchecksum\x18\x01 \x01(\fR\bchecksum\"4\n" +
+	"\bResponse\x12\x0e\n" +
+	"\x02ok\x18\x01 \x01(\bR\x02ok\x12\x18\n" +
+	"\amessage\x18\x02 \x01(\tR\amessage\"F\n" +
+	"\x11RetrievalResponse\x12\x1d\n" +
+	"\x04resp\x18\x01 \x01(\v2\t.ResponseR\x04resp\x12\x12\n" +
+	"\x04size\x18\x02 \x01(\x04R\x04size\"\x99\x02\n" +
+	"\aWrapper\x12'\n" +
+	"\bresponse\x18\x01 \x01(\v2\t.ResponseH\x00R\bresponse\x122\n" +
+	"\vstorage_req\x18\x02 \x01(\v2\x0f.StorageRequestH\x00R\n" +
+	"storageReq\x128\n" +
+	"\rretrieval_req\x18\x03 \x01(\v2\x11.RetrievalRequestH\x00R\fretrievalReq\x12;\n" +
+	"\x0eretrieval_resp\x18\x04 \x01(\v2\x12.RetrievalResponseH\x00R\rretrievalResp\x123\n" +
+	"\bchecksum\x18\x05 \x01(\v2\x15.ChecksumVerificationH\x00R\bchecksumB\x05\n" +
+	"\x03msgB\fZ\n" +
+	"./messagesb\x06proto3"
 
 var (
 	file_messages_proto_rawDescOnce sync.Once
-	file_messages_proto_rawDescData = file_messages_proto_rawDesc
+	file_messages_proto_rawDescData []byte
 )
 
 func file_messages_proto_rawDescGZIP() []byte {
 	file_messages_proto_rawDescOnce.Do(func() {
-		file_messages_proto_rawDescData = protoimpl.X.CompressGZIP(file_messages_proto_rawDescData)
+		file_messages_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_messages_proto_rawDesc), len(file_messages_proto_rawDesc)))
 	})
 	return file_messages_proto_rawDescData
 }
 
 var file_messages_proto_msgTypes = make([]protoimpl.MessageInfo, 6)
-var file_messages_proto_goTypes = []interface{}{
+var file_messages_proto_goTypes = []any{
 	(*StorageRequest)(nil),       // 0: StorageRequest
 	(*RetrievalRequest)(nil),     // 1: RetrievalRequest
 	(*ChecksumVerification)(nil), // 2: ChecksumVerification
@@ -486,81 +463,7 @@ func file_messages_proto_init() {
 	if File_messages_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_messages_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*StorageRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_messages_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RetrievalRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_messages_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*ChecksumVerification); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_messages_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Response); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_messages_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RetrievalResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_messages_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Wrapper); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
-	file_messages_proto_msgTypes[5].OneofWrappers = []interface{}{
+	file_messages_proto_msgTypes[5].OneofWrappers = []any{
 		(*Wrapper_Response)(nil),
 		(*Wrapper_StorageReq)(nil),
 		(*Wrapper_RetrievalReq)(nil),
@@ -571,7 +474,7 @@ func file_messages_proto_init() {
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_messages_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_messages_proto_rawDesc), len(file_messages_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   6,
 			NumExtensions: 0,
@@ -582,7 +485,6 @@ func file_messages_proto_init() {
 		MessageInfos:      file_messages_proto_msgTypes,
 	}.Build()
 	File_messages_proto = out.File
-	file_messages_proto_rawDesc = nil
 	file_messages_proto_goTypes = nil
 	file_messages_proto_depIdxs = nil
 }
